pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nonebot-plugin-osu-match-monitor"
version = "1.0.0"
description = "A NoneBot2 plugin used to monitor osu! match activities using match id."
readme = "README.md"
keywords = ["nonebot2", "osu!", "osu"]
license = { file = "LICENSE" }
authors = [
    { name = "Sevenyine", email = "tntobsidian@126.com" }
]
classifiers = [
    "Framework :: AsyncIO",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-adapter-onebot>=2.0.0",
    "httpx>=0.23.0"
]

[project.urls]
Homepage = "https://github.com/Sevenyine/nonebot-plugin-osu-match-monitor"
Documentation = "https://github.com/Sevenyine/nonebot-plugin-osu-match-monitor/blob/master/README.md"
Source = "https://github.com/Sevenyine/nonebot-plugin-osu-match-monitor"
Tracker = "https://github.com/Sevenyine/nonebot-plugin-osu-match-monitor/issues"